python/requirements.txt
websocket-client
requests
colorama